Moves transform tests to separate folder
