+++ test/transforms/new-page.js
@@ -2,11 +2,11 @@ const fs = require('fs');
 const path = require('path');
 const test = require('ava');
 
-const renameImportTransform = require('../source/transforms/rename-import-json');
-const renameJSXTransform = require('../source/transforms/rename-jsx');
+const renameImportTransform = require('../../source/transforms/rename-import-json');
+const renameJSXTransform = require('../../source/transforms/rename-jsx');
 
 const templateSource = fs.readFileSync(
-  path.join(__dirname, '../source/templates/mockup-page.jsx'),
+  path.join(__dirname, '../../source/templates/mockup-page.jsx'),
   'utf-8'
 );
