+++ test/transforms/snapshots/new-page.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/new-page-transform.js`
+# Snapshot report for `test/transforms/new-page.js`
 
-The actual snapshot is saved in `new-page-transform.js.snap`.
+The actual snapshot is saved in `new-page.js.snap`.
 
 Generated by [AVA](https://ava.li).
