+++ test/transforms/snapshots/rename.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/rename-transform.js`
+# Snapshot report for `test/transforms/rename.js`
 
-The actual snapshot is saved in `rename-transform.js.snap`.
+The actual snapshot is saved in `rename.js.snap`.
 
 Generated by [AVA](https://ava.li).
