+++ test/transforms/snapshots/to-stateful.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/stateful-transform.js`
+# Snapshot report for `test/transforms/to-stateful.js`
 
-The actual snapshot is saved in `stateful-transform.js.snap`.
+The actual snapshot is saved in `to-stateful.js.snap`.
 
 Generated by [AVA](https://ava.li).
