+++ test/transforms/snapshots/to-stateless.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/stateless-transform.js`
+# Snapshot report for `test/transforms/to-stateless.js`
 
-The actual snapshot is saved in `stateless-transform.js.snap`.
+The actual snapshot is saved in `to-stateless.js.snap`.
 
 Generated by [AVA](https://ava.li).
