+++ test/transforms/to-stateful.js
@@ -2,12 +2,12 @@ const fs = require('fs');
 const path = require('path');
 const test = require('ava');
 
-const statefulTransform = require('../source/transforms/to-stateful');
+const statefulTransform = require('../../source/transforms/to-stateful');
 
 const statelessComponentSource = fs.readFileSync(
   path.join(
     __dirname,
-    '../test-components/component-stateless/component-stateless.jsx'
+    '../../test-components/component-stateless/component-stateless.jsx'
   ),
   'utf-8'
 );
