+++ test/transforms/to-stateless.js
@@ -2,12 +2,12 @@ const fs = require('fs');
 const path = require('path');
 const test = require('ava');
 
-const statelessTransform = require('../source/transforms/to-stateless');
+const statelessTransform = require('../../source/transforms/to-stateless');
 
 const statefulComponentSource = fs.readFileSync(
   path.join(
     __dirname,
-    '../test-components/component-stateful/component-stateful.jsx'
+    '../../test-components/component-stateful/component-stateful.jsx'
   ),
   'utf-8'
 );
